+++ file.c
@@ -13,28 +13,28 @@ FILE * readFile(const char * filename) {
 
 char * readLine(FILE * stream) {
     char * output;
-    output = (char *) malloc(USHRT_MAX);
+    output = (char *) malloc(UINT16_MAX);
     register char currentChar;
-    register short index = 0;
+    register uint16_t index = 0;
     do {
         currentChar = fgetc(stream);
         if (currentChar == EOF) {
             return NULL;
         }
         output[index++] = currentChar;
-    } while (currentChar != '\n' && index < USHRT_MAX);
+    } while (currentChar != '\n' && index < UINT16_MAX);
     return output;
 }
 
 void trim(char * str) {
-    size_t i;
+    uint16_t i;
     for (i = 0; str[i] != '\0'; i++) {
         // isn't a whitespace (space or tab)
         if (str[i] != 32 && str[i] != 9) {
             break;
         }
     }
-    for (size_t j = 0; ; j++) {
+    for (uint16_t j = 0; ; j++) {
         str[j] = str[j + i];
         if (str[j] == '\0') {
             break;
