+++ type.h
@@ -2,18 +2,18 @@
 #define _TYPE_H_
 #include <stdio.h>
 #include <string.h>
+#include <stdint.h>
 #include <limits.h>
 #include <time.h>
-//#include "config.h"
 #include "color.h"
 
-extern long time_delay;
+extern uint32_t time_delay;
 
 void type(const char * text);
 
 void cursor();
 
-void delay(long milliseconds);
+void delay(uint32_t milliseconds);
 
 void sound();
